backend/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.4.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-jose>=3.3.0
email-validator>=2.1.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0  # For async HTTP client